#9220
Fixed the same class of debugWrite()/debugStart() printf conversion
mismatch #9115 fixed in the protocol modules, this time in the auths,
triggers, querytranslations, errortranslations, and directives modules:
a %s against a uint32_t salt in the postgresql auth modules, several
%.*s precision arguments narrower than the size_t/ssize_t they were
given, a %d against an int64_t errnum, a %lld against a uint64_t
affected-rows count with no cast, a %lld against an int64_t error
number with no cast, and a %s/%lld mismatch in a directive module.

Found by compiling every non-protocol module directory against a
scratch copy of sqlrserver.h with __attribute__((format(printf,...)))
added to debugStart()/debugWrite(), the same technique #9115 used.
filters, loggers, and the translation/router/query/moduledata module
types, plus src/server, were already clean.

Added the format attribute for real this time, as SQLR_PRINTF_FORMAT
(guarded on __GNUC__) on the real debugStart()/debugWrite()
declarations in sqlrserver.h, so the compiler catches the next one of
these at compile time. Fixed the one mismatch that turned up under a
full -Werror build with the real attribute in place: five zero-length
debugWrite(""); calls in the teradata protocol module, changed to
debugWrite("%s","");.

+++ src/errortranslations/patterns.cpp
@@ -301,9 +301,9 @@ void sqlrerrortranslation_patterns::applyPattern(const char *str,
 
 		debugWrite("applying regex");
 		debugWrite("from:");
-		debugWrite("\"%.*s%s\"",pfromlen,p->from,fromellipses);
+		debugWrite("\"%.*s%s\"",(int)pfromlen,p->from,fromellipses);
 		debugWrite("to:");
-		debugWrite("\"%.*s%s\"",ptolen,p->to,toellipses);
+		debugWrite("\"%.*s%s\"",(int)ptolen,p->to,toellipses);
 
 		convstr=charstring::replace(str,
 					p->fromre,
@@ -315,9 +315,9 @@ void sqlrerrortranslation_patterns::applyPattern(const char *str,
 
 		debugWrite("applying string:");
 		debugWrite("from:");
-		debugWrite("\"%.*s%s\"",pfromlen,p->from,fromellipses);
+		debugWrite("\"%.*s%s\"",(int)pfromlen,p->from,fromellipses);
 		debugWrite("to:");
-		debugWrite("\"%.*s%s\"",ptolen,p->to,toellipses);
+		debugWrite("\"%.*s%s\"",(int)ptolen,p->to,toellipses);
 		if (p->scope==SCOPE_INSIDE_QUOTES) {
 			debugWrite("inside quotes on chunk:");
 			debugWrite("\"%s\"",str);
