+++ src/querytranslations/patterns.cpp
@@ -290,9 +290,9 @@ void sqlrquerytranslation_patterns::applyPattern(const char *str,
 	} else if (p->fromre) {
 
 		debugWrite("applying regex from:");
-		debugWrite("\"%.*s%s\"",pfromlen,p->from,fromellipses);
+		debugWrite("\"%.*s%s\"",(int)pfromlen,p->from,fromellipses);
 		debugWrite("to:");
-		debugWrite("\"%.*s%s\"",ptolen,p->to,toellipses);
+		debugWrite("\"%.*s%s\"",(int)ptolen,p->to,toellipses);
 
 		convstr=charstring::replace(str,
 					p->fromre,
@@ -303,9 +303,9 @@ void sqlrquerytranslation_patterns::applyPattern(const char *str,
 	} else {
 
 		debugWrite("applying string from:");
-		debugWrite("\"%.*s%s\"",pfromlen,p->from,fromellipses);
+		debugWrite("\"%.*s%s\"",(int)pfromlen,p->from,fromellipses);
 		debugWrite("to:");
-		debugWrite("\"%.*s%s\"",ptolen,p->to,toellipses);
+		debugWrite("\"%.*s%s\"",(int)ptolen,p->to,toellipses);
 		if (p->scope==SCOPE_INSIDE_QUOTES) {
 			debugWrite("inside quotes on chunk:");
 			debugWrite("\"%s\"",str);
